+++ Tests/StreamChatTests/Database/DTOs/MessageDTO_Tests.swift
@@ -98,6 +98,7 @@ class MessageDTO_Tests: XCTestCase {
         XCTAssertEqual(channelPayload.config.connectEventsEnabled, loadedChannel?.config.connectEventsEnabled)
         XCTAssertEqual(channelPayload.config.uploadsEnabled, loadedChannel?.config.uploadsEnabled)
         XCTAssertEqual(channelPayload.config.repliesEnabled, loadedChannel?.config.repliesEnabled)
+        XCTAssertEqual(channelPayload.config.quotesEnabled, loadedChannel?.config.quotesEnabled)
         XCTAssertEqual(channelPayload.config.searchEnabled, loadedChannel?.config.searchEnabled)
         XCTAssertEqual(channelPayload.config.mutesEnabled, loadedChannel?.config.mutesEnabled)
         XCTAssertEqual(channelPayload.config.urlEnrichmentEnabled, loadedChannel?.config.urlEnrichmentEnabled)